requirements.txt
# SystemGen app dependencies
Flask>=3.0.0
Flask-Login>=0.6.3
Flask-WTF>=1.2.1
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.0.5
SQLAlchemy>=2.0.32
WTForms>=3.1.2
email-validator>=2.1.1

# ASGI/WSGI server
gunicorn>=21.2.0

# LLM provider and tooling
openai>=1.35.10
langchain>=0.2.13
langchain-openai>=0.1.23
tiktoken>=0.7.0
httpx>=0.27.0

# Optional utilities
python-dotenv>=1.0.1

# File extraction for uploads (Office/PDF)
python-docx>=1.1.2
python-pptx>=1.0.2
openpyxl>=3.1.5
pdfminer.six>=20231228